// frontend/src/app/shared/models/mapping/mappingdata.ts
export class ImappingProposalsTableData {
    sequence_no: number;
    ideal_text: string;
    stc_id: string;
    language_code: string;
    language_Id: number;
    rating: string;
    brand_name: string;
    type: string;
    project: string;
    numerous: string;
    gender: number;
    description: string;
    brand_id: number;
    parent_stc_id: number;
}